major rewrite: WIP push
got A* working, pushing

+++ src/searches/astar_search.h
@@ -4,4 +4,4 @@
 #include <string>
 
 RetVal astar_search(int startx, int starty, int goalx, int goaly, int width, int height, const std::vector<std::string>& citymap);
-RetVal astar_search(int startx, int starty, int goalx, int goaly, int width, int height, const std::vector<int>& citymap);
+RetVal astar_search(int startx, int starty, int goalx, int goaly, int width, int height, const std::vector<uint8_t>& citymap);
